rename and move DisclosureWidget to BouquetGroup
